+++ src/main/java/agents/storage/arbitrageStrategists/ProfitMaximiserPriceTaker.java
@@ -1,10 +1,11 @@
-// SPDX-FileCopyrightText: 2024 German Aerospace Center <amiris@dlr.de>
+// SPDX-FileCopyrightText: 2025 German Aerospace Center <amiris@dlr.de>
 //
 // SPDX-License-Identifier: Apache-2.0
 package agents.storage.arbitrageStrategists;
 
+import agents.markets.meritOrder.Constants;
 import agents.markets.meritOrder.sensitivities.MeritOrderSensitivity;
-import agents.markets.meritOrder.sensitivities.PriceSensitivity;
+import agents.markets.meritOrder.sensitivities.PriceNoSensitivity;
 import agents.markets.meritOrder.sensitivities.StepPower;
 import agents.storage.Device;
 import de.dlr.gitlab.fame.agent.input.ParameterData;
@@ -12,20 +13,21 @@ import de.dlr.gitlab.fame.agent.input.ParameterData.MissingDataException;
 import de.dlr.gitlab.fame.time.TimePeriod;
 
 /** Strategy to maximise profits via dynamic programming, running backwards in time. Profits are maximised by finding the best
- * sequence of states. The result fully utilises market power to improve profits.
+ * sequence of states. In contrast to ProfitMaximiser, the optimisation does not account for potential impact on prices, but can
+ * be considered as a "price taker".
  * 
- * @author Christoph Schimeczek, Evelyn Sperber */
-public class ProfitMaximiser extends DynamicProgrammingStrategist {
+ * @author Felix Nitsch, Christoph Schimeczek */
+public class ProfitMaximiserPriceTaker extends DynamicProgrammingStrategist {
 	/** incomeSum[t][i]: income that can be collected in time step t being in internal state i */
 	private final double[][] incomeSum;
 
-	/** Creates a {@link ProfitMaximiser}
+	/** Creates a {@link ProfitMaximiserPriceTaker}
 	 * 
 	 * @param generalInput general parameters associated with strategists
 	 * @param specificInput specific parameters for this strategist
 	 * @param storage device to be optimised
 	 * @throws MissingDataException if any required input is missing */
-	public ProfitMaximiser(ParameterData generalInput, ParameterData specificInput, Device storage)
+	public ProfitMaximiserPriceTaker(ParameterData generalInput, ParameterData specificInput, Device storage)
 			throws MissingDataException {
 		super(generalInput, specificInput, storage);
 		incomeSum = new double[forecastSteps][numberOfEnergyStates];
@@ -44,21 +46,21 @@ public class ProfitMaximiser extends DynamicProgrammingStrategist {
 	/** update most profitable final state for each possible initial state in every period */
 	@Override
 	protected void optimiseDispatch(TimePeriod firstPeriod) {
+		double[] powerDeltasInMW = calcPowerSteps();
 		for (int k = 0; k < forecastSteps; k++) {
 			int period = forecastSteps - k - 1; // step backwards in time
 			int nextPeriod = period + 1;
 			TimePeriod timePeriod = firstPeriod.shiftByDuration(period);
-			double[] chargePrices = calcChargePrices(timePeriod);
-			double[] powerDeltasInMW = calcPowerDeltas(timePeriod);
-
+			double chargePrice = calcChargePrice(timePeriod);
 			for (int initialState = 0; initialState < numberOfEnergyStates; initialState++) {
 				double currentBestIncome = -Double.MAX_VALUE;
 				int bestFinalState = Integer.MIN_VALUE;
+
 				int firstFinalState = calcFinalStateLowerBound(initialState);
 				int lastFinalState = calcFinalStateUpperBound(initialState);
 				for (int finalState = firstFinalState; finalState <= lastFinalState; finalState++) {
 					int stateDelta = finalState - initialState;
-					double incomeTransition = calcIncomeTransition(stateDelta, chargePrices, powerDeltasInMW);
+					double incomeTransition = calcIncomeTransition(stateDelta, chargePrice, powerDeltasInMW);
 					double income = incomeTransition + getBestIncome(nextPeriod, finalState);
 					if (income > currentBestIncome) {
 						currentBestIncome = income;
@@ -74,34 +76,26 @@ public class ProfitMaximiser extends DynamicProgrammingStrategist {
 		}
 	}
 
-	/** @return price steps for charging & discharging in the specified {@link TimePeriod} */
-	private double[] calcChargePrices(TimePeriod timePeriod) {
-		final PriceSensitivity sensitivity = (PriceSensitivity) getSensitivityForPeriod(timePeriod);
-		if (sensitivity != null) {
-			return sensitivity.getValuesInSteps(numberOfTransitionStates);
-		} else {
-			return new double[2 * numberOfTransitionStates + 1];
-		}
-	}
-
-	/** @return powers from max discharging to max charging in the specified {@link TimePeriod} */
-	private double[] calcPowerDeltas(TimePeriod timePeriod) {
-		double[] powerDeltasInMW = new double[numberOfTransitionStates * 2 + 1];
-		StepPower stepPower = calcStepPower(timePeriod);
+	/** @return power steps from max discharging to max charging */
+	private double[] calcPowerSteps() {
+		StepPower stepPower = new StepPower(storage.getExternalChargingPowerInMW(),
+				storage.getExternalDischargingPowerInMW(), numberOfTransitionStates);
+		double[] externalPowerStepsInMW = new double[numberOfTransitionStates * 2 + 1];
 		for (int i = -numberOfTransitionStates; i <= numberOfTransitionStates; i++) {
 			int index = i + numberOfTransitionStates;
-			powerDeltasInMW[index] = stepPower.getPower(i);
+			externalPowerStepsInMW[index] = stepPower.getPower(i);
 		}
-		return powerDeltasInMW;
+		return externalPowerStepsInMW;
 	}
 
-	/** @return power steps for charging & discharging in the specified {@link TimePeriod} */
-	private StepPower calcStepPower(TimePeriod timePeriod) {
-		final PriceSensitivity sensitivity = (PriceSensitivity) getSensitivityForPeriod(timePeriod);
+	/** @return price for (dis-)charging in the specified {@link TimePeriod}; assumes price of Zero if no valid forecast exists */
+	private double calcChargePrice(TimePeriod timePeriod) {
+		final PriceNoSensitivity sensitivity = (PriceNoSensitivity) getSensitivityForPeriod(timePeriod);
 		if (sensitivity != null) {
-			return sensitivity.getStepPowers(numberOfTransitionStates);
+			double priceForecast = sensitivity.getPriceForecast();
+			return Double.isNaN(priceForecast) ? 0 : priceForecast;
 		} else {
-			return new StepPower(0, 0, numberOfTransitionStates);
+			return 0;
 		}
 	}
 
@@ -110,31 +104,26 @@ public class ProfitMaximiser extends DynamicProgrammingStrategist {
 		return period < forecastSteps ? incomeSum[period][state] : 0;
 	}
 
-	/** @return income for a state transition under specified chargePrices */
-	private double calcIncomeTransition(int stateDelta, double[] chargePrices, double[] powerDeltasInMW) {
+	/** @return income for a state transition under specified chargePrice */
+	private double calcIncomeTransition(int stateDelta, double chargePrice, double[] powerDeltasInMW) {
 		int arrayIndex = numberOfTransitionStates + stateDelta;
 		double externalEnergyDelta = powerDeltasInMW[arrayIndex];
-		double price = chargePrices[arrayIndex];
-		return -externalEnergyDelta * price;
+		return -externalEnergyDelta * chargePrice;
 	}
 
 	@Override
 	protected double calcBidPrice(TimePeriod timePeriod, double externalEnergyDelta) {
 		if (externalEnergyDelta == 0) {
 			return Double.NaN;
+		} else if (externalEnergyDelta < 0) {
+			return Constants.MINIMAL_PRICE_IN_EUR_PER_MWH;
 		} else {
-			final PriceSensitivity sensitivity = (PriceSensitivity) getSensitivityForPeriod(timePeriod);
-			return sensitivity.calcPriceForExternalEnergyDelta(externalEnergyDelta);
+			return Constants.SCARCITY_PRICE_IN_EUR_PER_MWH;
 		}
 	}
 
 	@Override
 	protected MeritOrderSensitivity createBlankSensitivity() {
-		return new PriceSensitivity();
-	}
-
-	@Override
-	public void storeElectricityPriceForecast(TimePeriod timePeriod, double electricityPriceForecastInEURperMWH) {
-		throw new RuntimeException(ERR_USE_PRICE_FORECAST + getClass().getSimpleName());
+		return new PriceNoSensitivity();
 	}
 }